Fix section anchors.
Previously if you navigated to a section anchor, it would take you to
right *below* the section heading.  This isn't what you want; you want
to know that you actually got to the proper section.

It's possible to solve this with the approach taken in
anchored-headings.html, but

  * we are parsing HTML using regular expressions: https://stackoverflow.com/a/1732454/61394
  * using JS for this seems to be the canonical way to do it in Hugo,
    see e.g. https://github.com/gohugoio/hugoDocs/blob/0373e4310263ac9f2b0307d437dcf69cf474eb69/themes/gohugoioTheme/src/js/anchorforid.js

Instead, simply use anchorjs.

Note that this changes the anchor icon from `#` to `🔗`.  I can change
it back to `#` if desired, or we can make it configurable.

Doing what could be server-side rendering on the client makes me feel a
little icky, but it's not much different than how we handle mathjax or
syntax highlighting.